dm8_excel_etl/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dm8-excel-etl"
version = "0.1.0"
description = "Offline Excel -> DM8 (ODBC) ETL tool for Kunpeng + KylinOS."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "pyodbc==5.1.0",
  "openpyxl==3.1.5",
  "PyYAML==6.0.2",
  "python-dateutil==2.9.0.post0",
]

[project.optional-dependencies]
jdbc = [
  "jaydebeapi==1.2.3",
  "JPype1==1.5.0",
]

[project.scripts]
dm8-etl = "etl.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
